Implemented Presentation layer manager resource kpi service

+++ src/Application/ValueObjects/Accounts/UserKpi.cs
@@ -3,8 +3,8 @@ using ReSR.Application.ValueObjects.Core;
 namespace ReSR.Application.ValueObjects.Accounts;
 public readonly record struct UserKpi(
     DateRange DateRange,
-    int Count,
-    int FriendshipCount,
-    int BookmarkCount,
-    int ResourceCount
+    long Count,
+    long FriendshipCount,
+    long BookmarkCount,
+    long ResourceCount
 );
